UP-3235: OTA deployment of PRIOS

+++ cookbooks/deploy/recipes/jspr.rb
@@ -6,9 +6,14 @@ bash 'Deploying websites' do
     rm --recursive --force /var/www/Compass
     cp -r #{node['deploy_scripts_dir']}/JSPR/* /var/www
     ln -s /var/www/JSPR /var/www/Compass/JSPR
+  EOF
+end
 
+bash 'Deploying websites' do
+  code <<-EOF
     cp -r #{node['deploy_scripts_dir']}/Prios/* /var/www/Compass
   EOF
+  only_if { File.exists?('/var/www/Compass/Prios.ipa') }
 end
 
 template '/var/www/Compass/settings.js' do
